+++ src/main/java/kimp/websocket/client/BinanceWebSocketClient.java
@@ -3,7 +3,7 @@ package kimp.websocket.client;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import kimp.websocket.dto.response.BinanceReceiveDto;
 import kimp.websocket.dto.response.BinanceStreamDto;
-import kimp.websocket.handler.BinanceWebsocketHandler;
+import kimp.market.handler.BinanceWebsocketHandler;
 import lombok.extern.slf4j.Slf4j;
 import org.java_websocket.client.WebSocketClient;
 import org.java_websocket.handshake.ServerHandshake;
